Rename `test_fixtures` to simple `fixtures`.
